refactor: bulk store
