+++ src/extension/app/plugins/delete/delete.js
@@ -15,7 +15,6 @@ import {
   MODALS,
   MODAL_EVENTS,
   RESTRICTED_PATHS,
-  STATE,
 } from '../../constants.js';
 import { newTab } from '../../utils/browser.js';
 
@@ -62,8 +61,6 @@ export function createDeletePlugin(appStore) {
         });
         modal.addEventListener(MODAL_EVENTS.CONFIRM, async () => {
           // perform delete
-          appStore.setState(STATE.DELETING);
-
           const res = await appStore.delete();
           if (res) {
             const actionCallback = () => {
